teest(calculate-cost): in-source
